@@ -97,26 +91,29 @@
     "rollup": "3.9.1",
     "rollup-plugin-delete": "^2.0.0",
     "rollup-plugin-dts": "^5.1.1",
-    "rollup-plugin-import-css": "^3.1.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
-    "rollup-plugin-terser": "^7.0.2",
-    "storybook-addon-designs": "^6.3.1",
-    "storybook-addon-pseudo-states": "^1.14.0",
+    "storybook": "^7.5.3",
+    "storybook-addon-pseudo-states": "^2.1.2",
     "styled-components": "^5.2.0",
-    "ts-jest": "^26.5.6",
     "ts-loader": "^9.1.2",
     "typedoc": "^0.22.12",
-    "typescript": "4.9.5",
-    "url-loader": "^4.1.0"
+    "typescript": "^5.0.2",
+    "url-loader": "^4.1.0",
+    "vite-plugin-dts": "^3.6.3",
+    "vite-plugin-svgr": "^4.1.0",
+    "vite-tsconfig-paths": "^4.2.1",
+    "vitest": "^0.34.6"
   },
   "dependencies": {
-    "@storybook/testing-library": "^0.2.2",
     "@tippyjs/react": "^4.2.5",
     "dayjs": "^1.8.34",
     "emotion-reset": "^3.0.1",
+    "eslint-plugin-unused-imports": "^3.0.0",
     "lodash": "^4.17.19",
+    "patch-package": "^8.0.0",
     "pluralize": "^8.0.0",
     "polished": "^3.4.4",
+    "postinstall-postinstall": "^2.1.0",
     "react-aria": "^3.26.0",
     "react-aria-components": "^1.0.0-alpha.5",
     "react-fast-compare": "^3.2.0",
