@@ -2,8 +2,8 @@ import { render, screen } from '../../test';
 import DropdownButton from './DropdownButton';
 import userEvent from '@testing-library/user-event';
 
-const mockOnButtonClick = jest.fn();
-const mockonOptionSelect = jest.fn();
+const mockOnButtonClick = vi.fn();
+const mockonOptionSelect = vi.fn();
 
 const renderComponent = () => {
   const props = {
@@ -17,39 +17,39 @@ const renderComponent = () => {
 
 describe('DropdownButton:', () => {
   afterEach(() => {
-    jest.clearAllMocks();
+    vi.clearAllMocks();
   });
 
-  it('should display a dropdown menu when the triangleDown icon is clicked', () => {
+  it('should display a dropdown menu when the triangleDown icon is clicked', async () => {
     renderComponent();
 
     const iconButton = screen.getByTestId('dropdown-toggle-icon-button');
 
-    userEvent.click(iconButton);
+    await userEvent.click(iconButton);
 
     expect(screen.getByTestId('dropdown-button-options_list')).toBeInTheDocument();
   });
 
-  it('should trigger the Button CTA when the button is clicked', () => {
+  it('should trigger the Button CTA when the button is clicked', async () => {
     renderComponent();
 
     const button = screen.getByTestId('dropdownbutton');
 
-    userEvent.click(button);
+    await userEvent.click(button);
 
     expect(mockOnButtonClick).toHaveBeenCalledTimes(1);
   });
 
-  it('should trigger the Option CTA when one of the options is clicked', () => {
+  it('should trigger the Option CTA when one of the options is clicked', async () => {
     renderComponent();
 
     const iconButton = screen.getByTestId('dropdown-toggle-icon-button');
 
-    userEvent.click(iconButton);
+    await userEvent.click(iconButton);
 
     const option = screen.getByTestId('ictinus_list_item_Item_1');
 
-    userEvent.click(option);
+    await userEvent.click(option);
 
     expect(mockonOptionSelect).toHaveBeenCalledTimes(1);
   });
