@@ -1,14 +1,14 @@
 import userEvent from '@testing-library/user-event';
 
-import { render, screen, selectDropdownOption, waitFor } from '../../test';
+import { render, screen, selectDropdownOption, sleep, waitFor } from '../../test';
 import StatefulSelect from './StatefulSelect';
 import { fireEvent } from '@testing-library/react';
 import { SELECT_ALL_OPTION } from './constants';
 
-global.ResizeObserver = jest.fn().mockImplementation(() => ({
-  observe: jest.fn(),
-  unobserve: jest.fn(),
-  disconnect: jest.fn(),
+global.ResizeObserver = vi.fn().mockImplementation(() => ({
+  observe: vi.fn(),
+  unobserve: vi.fn(),
+  disconnect: vi.fn(),
 }));
 
 const dropdownList = [
@@ -26,7 +26,7 @@ const dropdownListWithHelperText = [
 
 describe('Generic Select', () => {
   describe('Sync Select', () => {
-    const handleSubmit = jest.fn();
+    const handleSubmit = vi.fn();
 
     let selectInput: HTMLInputElement;
     let clearIcon: HTMLElement;
@@ -44,7 +44,7 @@ describe('Generic Select', () => {
     });
 
     afterEach(() => {
-      jest.clearAllMocks();
+      vi.clearAllMocks();
     });
 
     it('should change the input value correctly', async () => {
@@ -63,31 +63,31 @@ describe('Generic Select', () => {
 
     it('should revert input value back to “default”, when clear icon is clicked', async () => {
       selectInput = screen.getByPlaceholderText('Country') as HTMLInputElement;
-      userEvent.type(selectInput, 'Greece');
+      await userEvent.type(selectInput, 'Greece');
 
       expect(selectInput.value).toBe('Greece');
 
       clearIcon = screen.getByTestId('select-right-icon');
-      userEvent.click(clearIcon);
+      await userEvent.click(clearIcon);
 
       expect(selectInput.value).not.toBe('Greece');
     });
 
     it('should revert input value back to “default”, when Backspace is clicked', async () => {
       selectInput = screen.getByPlaceholderText('Country') as HTMLInputElement;
-      userEvent.type(selectInput, 'Greece');
+      await userEvent.type(selectInput, 'Greece');
 
       expect(selectInput.value).toBe('Greece');
 
-      userEvent.type(selectInput, '{Backspace}');
+      await userEvent.type(selectInput, '{Backspace}');
 
       expect(selectInput.value).not.toBe('Greece');
     });
   });
 
   describe('Async Select', () => {
-    const handleSubmit = jest.fn();
-    const asyncSearch = jest.fn();
+    const handleSubmit = vi.fn();
+    const asyncSearch = vi.fn();
 
     let selectInput: HTMLInputElement;
 
@@ -108,7 +108,7 @@ describe('Generic Select', () => {
     };
 
     afterEach(() => {
-      jest.clearAllMocks();
+      vi.clearAllMocks();
     });
 
     it('should display progress indicator when isLoading is true', async () => {
@@ -116,9 +116,13 @@ describe('Generic Select', () => {
 
       selectInput = screen.getByPlaceholderText('Country') as HTMLInputElement;
 
-      userEvent.type(selectInput, 'Greece');
+      await userEvent.type(selectInput, 'Greece');
 
-      await waitFor(() => expect(screen.getByTestId('select_circular_progress_container')).toBeVisible());
+      await sleep(500);
+
+      await waitFor(() =>
+        expect(screen.getByTestId('select_circular_progress_container')).toBeVisible()
+      );
     });
 
     it('should call asyncSearch when typing', async () => {
@@ -142,7 +146,7 @@ describe('Generic Select', () => {
   });
 
   describe('Select helper text option', () => {
-    const handleSubmit = jest.fn();
+    const handleSubmit = vi.fn();
 
     let selectInput: HTMLInputElement;
 
@@ -160,7 +164,7 @@ describe('Generic Select', () => {
     };
 
     afterEach(() => {
-      jest.clearAllMocks();
+      vi.clearAllMocks();
     });
 
     it('should display helper text when passed as prop', async () => {
@@ -168,7 +172,7 @@ describe('Generic Select', () => {
 
       selectInput = screen.getByPlaceholderText('Countries') as HTMLInputElement;
 
-      userEvent.click(selectInput);
+      await userEvent.click(selectInput);
 
       await waitFor(() => expect(screen.getByText('Greece')).toBeVisible());
       await waitFor(() => expect(screen.getByText('Europe')).toBeVisible());
@@ -176,6 +180,10 @@ describe('Generic Select', () => {
   });
 
   describe('Keyboard Navigation', () => {
+    afterEach(() => {
+      // reset mocked time
+      vi.useRealTimers();
+    });
     it('down arrow should open menu', async () => {
       const { getByTestId, container } = render(
         <div>
@@ -253,8 +261,7 @@ describe('Generic Select', () => {
     });
 
     it('on type more than 1 character in input and enter selects the first from the list', async () => {
-      jest.useFakeTimers();
-      const handleSubmit = jest.fn();
+      const handleSubmit = vi.fn();
 
       const { getByTestId } = render(
         <div>
@@ -280,11 +287,8 @@ describe('Generic Select', () => {
       expect(handleSubmit).toBeCalledTimes(1);
 
       await userEvent.type(select, 'zi');
-      fireEvent.keyDown(select, {
-        key: 'Enter',
-      });
 
-      jest.runAllTimers();
+      await userEvent.keyboard('{enter}');
 
       expect(handleSubmit).toBeCalledTimes(2);
       expect(handleSubmit).toBeCalledWith(dropdownList[1]);
@@ -317,7 +321,7 @@ describe('Multi Select', () => {
   });
 
   afterEach(() => {
-    jest.clearAllMocks();
+    vi.clearAllMocks();
   });
 
   it('renders the Chips when options are clicked', async () => {
@@ -332,7 +336,7 @@ describe('Multi Select', () => {
     await selectDropdownOption(selectInput, dropdownList[0].label);
     await selectDropdownOption(selectInput, dropdownList[1].label);
 
-    userEvent.click(selectInput);
+    await userEvent.click(selectInput);
 
     expect(screen.getByTestId('ictinus_list').innerHTML).not.toContain(dropdownList[0].label);
     expect(screen.getByTestId('ictinus_list').innerHTML).not.toContain(dropdownList[1].label);
@@ -342,10 +346,10 @@ describe('Multi Select', () => {
     await selectDropdownOption(selectInput, dropdownList[0].label);
     await selectDropdownOption(selectInput, dropdownList[1].label);
 
-    userEvent.click(screen.getByTestId('chip-delete-chip_1'));
-    userEvent.click(screen.getByTestId('chip-delete-chip_0'));
+    await userEvent.click(screen.getByTestId('chip-delete-chip_1'));
+    await userEvent.click(screen.getByTestId('chip-delete-chip_0'));
 
-    userEvent.click(selectInput);
+    await userEvent.click(selectInput);
 
     expect(screen.getByTestId('ictinus_list').innerHTML).toContain(dropdownList[0].label);
     expect(screen.getByTestId('ictinus_list').innerHTML).toContain(dropdownList[1].label);
@@ -356,40 +360,40 @@ describe('Multi Select', () => {
     await selectDropdownOption(selectInput, dropdownList[1].label);
     await selectDropdownOption(selectInput, dropdownList[2].label);
 
-    userEvent.click(screen.getByTestId('select-right-icon'));
+    await userEvent.click(screen.getByTestId('select-right-icon'));
 
     expect(screen.queryByTestId('chip-chip_0')).not.toBeInTheDocument();
   });
 
   it('selects all options when Select All is clicked', async () => {
-    userEvent.click(selectInput);
-    userEvent.click(screen.getByTestId(`ictinus_list_item_${SELECT_ALL_OPTION.value}`));
+    await userEvent.click(selectInput);
+    await userEvent.click(screen.getByTestId(`ictinus_list_item_${SELECT_ALL_OPTION.value}`));
 
     chips = await screen.findAllByTestId(/chip-chip_/);
 
     expect(chips.length).toEqual(dropdownList.length);
 
-    userEvent.click(selectInput);
+    await userEvent.click(selectInput);
 
     expect(screen.getByText('No options')).toBeInTheDocument();
   });
 
   it('creates a new item if not found on the results', async () => {
-    userEvent.click(selectInput);
-    userEvent.type(selectInput, 'New item');
+    await userEvent.click(selectInput);
+    await userEvent.type(selectInput, 'New item');
 
     expect(screen.getByTestId('ictinus_list').innerHTML).toContain('Create "New item"');
 
     /** Create new Item */
     newOption = screen.getByText('Create "New item"');
-    userEvent.click(newOption);
+    await userEvent.click(newOption);
     newChip = screen.getByTestId('chip-chip_0');
 
     expect(newChip).toBeVisible();
     expect(newChip.innerHTML).toContain('New item');
 
     /** Delete created Item */
-    userEvent.click(screen.getByTestId('chip-delete-chip_0'));
+    await userEvent.click(screen.getByTestId('chip-delete-chip_0'));
     expect(screen.queryByTestId('New item')).not.toBeInTheDocument();
   });
 
@@ -400,8 +404,8 @@ describe('Multi Select', () => {
     expect(screen.getByTestId('chip-chip_0')).toBeVisible();
     expect(screen.getByTestId('chip-chip_1')).toBeVisible();
 
-    userEvent.click(selectInput);
-    userEvent.type(selectInput, '{Backspace}');
+    await userEvent.click(selectInput);
+    await userEvent.type(selectInput, '{Backspace}');
 
     expect(screen.queryByTestId('chip-chip_1')).not.toBeInTheDocument();
   });
